requirements.txt
selenium==4.21.0
webdriver-manager==4.0.1
google-api-python-client==2.131.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
